requirements.txt
streamlit
langchain
langchain_groq
requests
python-dotenv
